+++ src/sentry_options.c
@@ -141,6 +141,14 @@ sentry_options_set_on_crash(
 }
 
 void
+sentry_options_set_before_breadcrumb(
+    sentry_options_t *opts, sentry_breadcrumb_function_t func, void *data)
+{
+    opts->on_breadcrumb_func = func;
+    opts->on_breadcrumb_data = data;
+}
+
+void
 sentry_options_set_dsn_n(
     sentry_options_t *opts, const char *raw_dsn, size_t raw_dsn_len)
 {