+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Features**:
 
 - Provide Clang-CL support. ([#1161](https://github.com/getsentry/sentry-native/pull/1161), [crashpad#100](https://github.com/getsentry/crashpad/pull/100))
+- Add `before_breadcrumb` hook ([#1166](https://github.com/getsentry/sentry-native/pull/1166))
 
 **Thank you**: